+++ UserInterface.Core/Services/FileService.cs
@@ -13,15 +13,11 @@
 
 
 
-#region
-
 using System.Text;
 
 using KC.WindowsConfigurationAnalyzer.UserInterface.Core.Contracts.Services;
 
 using Newtonsoft.Json;
-
-#endregion
 
 
 
@@ -36,14 +32,11 @@ public class FileService : IFileService
 
     public T? Read<T>(string folderPath, string fileName)
     {
-        string path = Path.Combine(folderPath, fileName);
+        var path = Path.Combine(folderPath, fileName);
 
-        if (!File.Exists(path))
-        {
-            return default;
-        }
+        if (!File.Exists(path)) return default;
 
-        string json = File.ReadAllText(path);
+        var json = File.ReadAllText(path);
 
         return JsonConvert.DeserializeObject<T>(json);
     }
@@ -54,12 +47,9 @@ public class FileService : IFileService
 
     public void Save<T>(string folderPath, string fileName, T content)
     {
-        if (!Directory.Exists(folderPath))
-        {
-            Directory.CreateDirectory(folderPath);
-        }
+        if (!Directory.Exists(folderPath)) Directory.CreateDirectory(folderPath);
 
-        string fileContent = JsonConvert.SerializeObject(content);
+        var fileContent = JsonConvert.SerializeObject(content);
         File.WriteAllText(Path.Combine(folderPath, fileName), fileContent, Encoding.UTF8);
     }
 
@@ -69,10 +59,7 @@ public class FileService : IFileService
 
     public void Delete(string folderPath, string fileName)
     {
-        if (fileName != null && File.Exists(Path.Combine(folderPath, fileName)))
-        {
-            File.Delete(Path.Combine(folderPath, fileName));
-        }
+        if (fileName != null && File.Exists(Path.Combine(folderPath, fileName))) File.Delete(Path.Combine(folderPath, fileName));
     }
